# docs/swagger.yaml
basePath: /
definitions:
...
  ds.EmissionCompleteRequest:
    properties:
      moderator_id:
        type: integer
      status:
        enum:
        - завершен
        - отклонен
        type: string
    required:
    - moderator_id
    - status
    type: object
  ds.EmissionListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/ds.EmissionResponse'
        type: array
      status:
        example: success
        type: string
    type: object
  ds.EmissionResponse:
    properties:
      create_at:
        type: string
      distance:
        type: number
      finish_at:
        type: string
      id:
        type: integer
      status:
        type: string
      total_emission:
        type: number
      update_at:
        type: string
    type: object
  ds.EmissionUpdateRequest:
    properties:
      distance:
        type: number
    type: object
  ds.ErrorResponse:
    properties:
      error:
        example: Некорректный токен
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API сервиса аутентификации с Redis и JWT. Поддерживает роли пользователей
    и администраторов.
  title: Device emission Service API
  version: "1.0"
paths:
...
  /api/emissions_calculation:
    get:
      consumes:
      - application/json
      description: 'Возвращает список заказов с возможностью фильтрации по статусу
        и диапазону дат. Разрешённые статусы: "сформирован", "завершен", "отклонен".'
      parameters:
      - description: Статус заказа (сформирован, завершен, отклонен), можно указать
          несколько через запятую
        in: query
        name: status
        type: string
      - description: Дата начала фильтрации (формат YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: Дата окончания фильтрации (формат YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком заказов
          schema:
            $ref: '#/definitions/ds.EmissionListResponse'
        "500":
          description: Ошибка на сервере
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список заказов
      tags:
      - Заявки с устройствами
  /api/emissions_calculation/:id:
    get:
      description: Возвращает полную информацию о заказе с расчетом выбросов, включая
        список устройств. Доступен только владельцу заказа или администратору
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с данными заказа и устройств
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID заказа
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (не владелец и не администратор)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заказ не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить заказ с устройствами
      tags:
      - Заявки с устройствами
    put:
      consumes:
      - application/json
      description: Обновляет параметры заказа (например, расстояние). Разрешено обновлять
        только поле distance
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления заказа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.EmissionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно обновлен
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID или недопустимые поля в запросе
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обновлении заказа
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить заказ
      tags:
      - Заявки с устройствами
  /api/emissions_calculation/:id/complete:
    put:
      consumes:
      - application/json
      description: Переводит заказ в статус "завершен" или "отклонен". Доступно только
        администраторам
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для завершения/отклонения заказа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.EmissionCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно обновлен
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID или данные запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль Admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обновлении заказа
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Завершить или отклонить заказ
      tags:
      - Заявки с устройствами
  /api/emissions_calculation/:id/form:
    put:
      description: Переводит черновой заказ в статус "сформирован". Требуется, чтобы
        все устройства имели заполненное поле custom_power
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно сформирован
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID заказа или не все custom_power заполнены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при формировании заказа
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Сформировать заказ
      tags:
      - Заявки с устройствами
...
  /api/emissions_calculation/delete/:id:
    post:
      description: Помечает заказ как удаленный (soft delete), устанавливая статус
        "удален"
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно удален
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID заказа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при удалении заказа
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить заказ
      tags:
      - Заявки с устройствами
...
  /api/emissions_calculation/draft/cart:
    get:
      description: 'Возвращает информацию о черновом заказе пользователя: ID заказа
        и количество устройств в нем'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с данными корзины
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка на сервере
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить корзину чернового заказа
      tags:
      - Заявки с устройствами
  /api/emissions_calculation/user:
    get:
      consumes:
      - application/json
      description: 'Возвращает список заказов с возможностью фильтрации по статусу
        и диапазону дат. Разрешённые статусы: "сформирован", "завершен", "отклонен".'
      parameters:
      - description: Статус заказа (сформирован, завершен, отклонен), можно указать
          несколько через запятую
        in: query
        name: status
        type: string
      - description: Дата начала фильтрации (формат YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: Дата окончания фильтрации (формат YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком заказов
          schema:
            $ref: '#/definitions/ds.EmissionListResponse'
        "500":
          description: Ошибка на сервере
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список заказов
      tags:
      - Заявки с устройствами
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"